Add basic support for libgpiod
The sysfs interface has been deprecated, add a very basic support for
using the new libgpiod. This library require a GPIO chip name which
can now be passed by adding @CHIPNAME after the GPIO index. The chip
name can then be ommited if the later GPIO are on the same chip.

The libpgiod support can be enabled:
  - by passing WITH_LIBGPIOD=1 to make.
  - with option '--enable-libgpiod' if you
  are using autotools scripts.

Signed-off-by: Alban Bedel <alban.bedel@aerq.com>
[Pierre-Jean:
  - add autotools support with libgpiod
  - rework commit log
]
Signed-off-by: Pierre-Jean Texier <pjtexier@koncepto.io>